+++ R/getTags.R
@@ -7,8 +7,19 @@
 #'
 #' @template debugTemplate
 #'
-#' @return [getTagMapping()] returns a data frame holding the `code` and
-#' `meaning` of each possible tag.
+#' @return [getTags()] returns a data frame holding the tags that
+#' have been saved for a particular data file.
+#'
+#' @examples
+#' # This example uses a temporary file and unlinks it later;
+#' # obviously, this is not what would be done in practice, but
+#' # it is necessary for the example to pass CRAN checks.
+#' dbname <- tempfile() # do not do this in practice
+#' createDatabase(dbname)
+#' saveTag("ctd.cnv", index = 1, tag = 2, analyst = "Dan Kelley", dbname = dbname)
+#' tags <- getTags("ctd.cnv", dbname = dbname)
+#' print(tags)
+#' unlink(dbname) # do not do this in practice
 #'
 #' @author Dan Kelley
 #'