+++ NEWS.md
@@ -1,6 +1,7 @@
 # ocetag 0.0.2
 
-* Minor retooling.
+* Change `createDatabase()` and `saveTag()` to handle arbitrary column names.
+* Add a vignette showing how to work with CTD data.
 
 # ocetag 0.0.1